Update durability; make trade_util2->trade_util
